requirements.txt
# HF - vLLM
transformers>=4.35.0
ninja  # For faster builds.
psutil
ray >= 2.5.1
numpy
torch == 2.1.2
einops
xformers == 0.0.23.post1  # Required for CUDA 12.1.
fastapi
uvicorn[standard]
pydantic==1.10.13
aioprometheus[starlette]
vllm
prometheus-fastapi-instrumentator

# TensorRT
accelerate==0.25.0
build
colored
diffusers==0.15.0
lark
mpi4py
numpy
onnx>=1.12.0
polygraphy
psutil
pynvml>=11.5.0
sentencepiece>=0.1.99
nvidia-ammo~=0.5.0; platform_machine=="x86_64"
transformers==4.36.1
wheel
optimum
evaluate
janus

# Quantization
autoawq==0.1.8
auto-gptq